# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "THE Question Portal",
        "description": "Create a Project using laravel and reactJS",
        "termsOfService": "",
        "contact": {
            "email": "sumersingh1997.ssh@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add-question": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Add Question",
                "operationId": "storeQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "text"
                                    },
                                    "answer": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/view-question/{id}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "View Question",
                "operationId": "showQuestion",
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/update-question-status/{id}": {
            "patch": {
                "tags": [
                    "Question"
                ],
                "summary": "Update Question Status",
                "operationId": "updateQuestionStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question Id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/update-question/{id}": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Update Question",
                "operationId": "updateQuestion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question Id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "text"
                                    },
                                    "answer": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/delete-question/{id}": {
            "delete": {
                "tags": [
                    "Question"
                ],
                "summary": "Delete Question",
                "operationId": "destroyQuestion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question Id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_security": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer < token >)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}